refactor(alpha-stack): rename to stack except directory name
